requirements.txt
anyio==4.4.0
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dnspython==1.16.0
fastapi==0.95.2
h11==0.14.0
idna==3.7
motor==3.5.1
pydantic==1.10.7
pymongo==4.8.0
python-dotenv==0.19.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
